More changes according comments
