+++ consoleApp/src/main/java/com/coherentsolutions/trainings/java/auto/consoleApp/ConfigReader.java
@@ -34,8 +34,4 @@ public class ConfigReader {
             throw new RuntimeException("Failed to read config file");
         }
     }
-
-    public static void main(String[] args) {
-        System.out.println(new ConfigReader().readOrder());
-    }
 }
