--- store/src/main/java/app/com/coherentsolutions/trainings/java/auto/store/Store2.java
@@ -1,13 +1,0 @@
-package app.com.coherentsolutions.trainings.java.auto.store;
-
-import java.util.LinkedList;
-import java.util.List;
-
-public class Store2 {
-    List<Category> categoryList;
-
-    public Store2() {
-        categoryList = new LinkedList<>();
-    }
-
-}
